requirements.txt
Django==1.11.10
django-filter==1.1.0
djangorestframework==3.7.7
pytz==2017.3
